requirements.txt
pandas
scikit-learn
dvc
mlflow
pytest
pyyaml
